ledger_service/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
httpx

# El driver oficial para conectarse a Cassandra
cassandra-driver